
\name{bigBinSize}
\alias{bigBinSize}
\docType{data}
\title{Big bin size}
\description{The deltaCaptureC package depends on rebinning mean
  normalized capture C data from a small bin size to a large bin size.
  This gives the size of the latter. This is used in package
  documentation. }
\usage{bigBinSize}
\format{a numeric scalar}
\keyword{datasets}