
\name{binnedDeltaPlot}
\alias{binnedDeltaPlot}
\docType{data}
\title{Plot of Binned Delta Counts}
\description{The data plotted here are the difference between mean
  normalized counts for captures in two replicates each of EScells and
  Neurons. The data here is binned to 1000bp and is shown in a region of
  500kb up- and down-stream of the Paupar viewpoint.  It is used here to
  illustrate functioning of the deltaCaptureC package.}
\usage{binnedDeltaPlot}
\format{A ggplot object}
\keyword{datasets}
