
\name{binnedDeltaSE}
\alias{binnedDeltaSE}
\docType{data}
\title{Binned difference of mean capture-C counts between EScells and Neurons}
\description{A smaller deltaSE restriction to a region of interest
  around the viewpoint, here binned to the bigBinSize, 10kb.}
\usage{binnedDeltaSE}
\format{A RangedSummarizedExperiment with 1909 rows.}
\keyword{datasets}