
\name{binnedSummarizedExperiment}
\alias{binnedSummarizedExperiment}
\docType{data}
\title{Binned Capture-C counts of EScells and Neurons}
\description{Capture-C counts of EScells and Neurons restricted to a
  region of interest around the Paupar viewpoint.  The data have been
  binned to smallBins, i.e., bins of size 1000.  Used in package
  documentation. }
\usage{binnedSummarizedExperiment}
\format{A RangedSummarizedExperiment.}
\keyword{datasets}