% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{binSummarizedExperiment}
\alias{binSummarizedExperiment}
\title{Bin a Summarized experiment into a set of bins given by a GRanges object}
\usage{
binSummarizedExperiment(bins, se, checkDisjoint = FALSE)
}
\arguments{
\item{bins}{a GRanges object whose ranges should be disjoint}

\item{se}{a RangedSummarizedExperiment}

\item{checkDisjoint}{= FALSE if set to true will check that the bins
are disjoint}
}
\value{
a RangedSummarizedExperiment
}
\description{
This function takes a set of bins given by a GRanges object and a
RangedSummarizedExperiment and produces a new RangedSummarizedExperiment
with the bins as its rowRanges
}
\examples{
binnedSummarizedExperiment = binSummarizedExperiment(smallSetOfSmallBins,smallerDeltaSE)
}
