
\name{deltaSE}
\alias{deltaSE}
\docType{data}
\title{Difference of mean capture-C counts between EScells and Neurons}
\description{This gives the difference in mean normalized capture C
  counts between two replicates each of EScells and Neurons.  It is the
  underlying kind of data on which deltaCaptureC operates and is here to
  support documentation.}
\usage{deltaSE}
\format{A RangedSummarizedExperiment with 5709 rows.}
\keyword{datasets}
