% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{.getRunsAndTotals}
\alias{.getRunsAndTotals}
\title{A helper function for getRunTotals}
\usage{
.getRunsAndTotals(gr, colName)
}
\arguments{
\item{gr}{a GRanges object whose mcols gives the relevant binned
data}

\item{colName}{This designates the column in mcols with the relevant
data}
}
\value{
a GRanges object giving the contiguous regions and their
respective sums.
}
\description{
This takes a GRanges object for binneed data and a column name
designating where to find the relevant data in the mcols and
returns a GRanges giving the consecutive runs of constant sign and
their run totals.  It is not exported.
}
