% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{downshiftDFtoMatrix}
\alias{downshiftDFtoMatrix}
\title{Downshift from DF to matrix}
\usage{
downshiftDFtoMatrix(df)
}
\arguments{
\item{df}{This is a data frame whose first three columns are chr, start and end and
whose remaining columns are numerical data}
}
\value{
A matrix of numerical data
}
\description{
This function takes a data.frame with chr, start, end and numerical data and turns
it into a matrix with row names chr:start-end
}
\examples{
m = downshiftDFtoMatrix(miniSEDF)
}
