% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{generatePermutation}
\alias{generatePermutation}
\title{Generate permutation for permutation testing}
\usage{
generatePermutation(gr, innerRegion)
}
\arguments{
\item{gr}{a GRanges object which should be ordered}

\item{innerRegion}{a GRanges object which should be a single interval}
}
\value{
a permutation of 1:length(gr)
}
\description{
This function takes a set of row ranges and an inner region and
generates a permutation which is symmetric on the inner region and
arbitrary on the remainder
}
\examples{
permutations = generatePermutation(smallBins,viewpointRegion)
}
