% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getDeltaSE}
\alias{getDeltaSE}
\title{Make delta summarized experiment:}
\usage{
getDeltaSE(countsSE, byTreatment = "treatment")
}
\arguments{
\item{countsSE}{A summarized experiment with assay counts and optionally assay normalized
counts}

\item{byTreatment}{= 'treatment' Allows for specifying some other condition than 'treatment'}
}
\value{
A summarized experiment with a single assay consisting of a single column, the
delta mean normalized counts.
}
\description{
This function takes a SummarizedExperiment with count data and produces
a SummarizedExperiment of the delta track.  There should exactly two values
for treatment, i.e., byTreatment
}
\examples{
aSmallDeltaSE = getDeltaSE(miniSE)
}
