% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getLopsidedness}
\alias{getLopsidedness}
\title{Get the lopsidedness statistic}
\usage{
getLopsidedness(se, viewpointRegion, colName = "delta")
}
\arguments{
\item{se}{a SummerizedExperiment giving the delta or permuted delta}

\item{viewpointRegion}{the region around the viewpoint in which to
investigate lopsidedness}

\item{colName}{defaults to 'delta'}
}
\value{
the lopsidedness around the viewpointMid in the
viewpointRegion
}
\description{
This function looks at the sidedness around the viewpoint and
returns the absolute value of the difference between the sum of
the values before and after the viewpoint inside the viewpoint
region.
}
\examples{
lopsidedness = getLopsidedness(binnedDeltaSE,viewpointRegion)
}
