% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getMeanNormalizedCountsSE}
\alias{getMeanNormalizedCountsSE}
\title{Make mean treatment summarized experiment:}
\usage{
getMeanNormalizedCountsSE(countsSE, byTreatment = "treatment")
}
\arguments{
\item{countsSE}{A SummarizedExperiment containing an assay 'counts' and
optionally an assay 'normalizedCounts'}

\item{byTreatment}{= 'treatment' This gives the column of colData to use for
taking averages}
}
\value{
A SummarizedExperiment giving mean normalized counts for each value of
byTreatment
}
\description{
Get the mean normalized counts for each treatment
}
\details{
This function takes a SummarizedExperiment. It looks for an assay called
normalizedCounts.  If this assay is missing, it creates it by normalizing
using the size factors. By default, it takes the mean for each value of
colData$treatment
}
\examples{
meanNormalizedCountSE = getMeanNormalizedCountsSE(miniSE)
}
