% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getNormalizedCountsSE}
\alias{getNormalizedCountsSE}
\title{Get normalized counts}
\usage{
getNormalizedCountsSE(se)
}
\arguments{
\item{se}{A SummarizedExperiment with an assay called counts giving
the raw counts for each replicate of the two treatments.}
}
\value{
A SummarizedExperiment including a an assay of the
normalized counts called normalizedCounts.
}
\description{
This function takes a SummarizedExperiment giving the the counts
for each replicate of the two treatments and computes and affixes
an assay giving the normalized version of these counts.
}
\examples{
miniSENormalized = getNormalizedCountsSE(miniSE)
}
