% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getRunAndLopsidednessStatistics}
\alias{getRunAndLopsidednessStatistics}
\title{Get the distribution of run and lopsidedness statistics}
\usage{
getRunAndLopsidednessStatistics(
  scrambledDeltas,
  viewpointRegion,
  colName = "delta"
)
}
\arguments{
\item{scrambledDeltas}{a list of rebinned (i.e., to large bin size)
of scrambled deltas}

\item{viewpointRegion}{a GRanges object giving the region that is
reserved for lopsidedness}

\item{colName}{= 'delta'}
}
\value{
a Nx4 matrix giving the min, max, max(abs(min),abs(max))
and lopsidedness for the run totals in the list of scrambled
deltas.
}
\description{
Get the distribution of run and lopsidedness statistics
}
