% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getRunStatistics}
\alias{getRunStatistics}
\title{This function is called by getRunsStatisticsDist on the individual elements
of a list of scrambled runs.}
\usage{
getRunStatistics(runTotals)
}
\arguments{
\item{runTotals}{is a GRanges object giving the consecutive runs
and their totals.}
}
\value{
a vector of the min, max and absolute value of the min and
max for the run totals.
}
\description{
This is a helper function.  Currently not exported.
}
