% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getRunStatisticsDist}
\alias{getRunStatisticsDist}
\title{This takes a list of (scrambled) runs and returns their run statistics}
\usage{
getRunStatisticsDist(runTotalsList)
}
\arguments{
\item{runTotalsList}{this is a list whose members are GRanges
objects giving the consecutive runs and their totals}
}
\value{
a Nx3 matrix giving the min, max and max(abs(min),abs(max))
run totals
}
\description{
This function takes a list of (scrambled) runs and extracts their
run totals as a matrix with colnames 'min','max' and 'abs', the
latter being the max of the absolute values of the previous two
}
