% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{getSizeFactorsDF}
\alias{getSizeFactorsDF}
\title{Get the size factors for count normalization}
\usage{
getSizeFactorsDF(countsDF)
}
\arguments{
\item{countsDF}{A data frame whose first three columns are chr, start and end, and
whose remaining columns are count data for experimental replicates}
}
\value{
The size factors for the columns of countsDF
}
\description{
This function takes a data frame giving chr, start, end and count for experimental
replicates and returns the size factors for each of the replicates for use in
normalization
}
\examples{
sf = getSizeFactorsDF(miniSEDF)
}
