
\name{miniDeltaSE}
\alias{miniDeltaSE}
\docType{data}
\title{Difference of mean capture-C counts between EScells and Neurons}
\description{A smaller deltaSE restriction to a region of interest
  around the viewpoint.}
\usage{miniDeltaSE}
\format{A RangedSummarizedExperiment with 1909 rows.}
\keyword{datasets}