
\name{miniSE}
\alias{miniSE}
\docType{data}
\title{Capture-C counts of EScells and Neurons}
\description{Capture-C counts of EScells and Neurons restricted to a
  region of interest around the Paupar viewpoint.  Used in package
  documentation. }
\usage{miniSE}
\format{A RangedSummarizedExperiment with 1909 rows.}
\keyword{datasets}