
\name{miniSEDF}
\alias{miniSEDF}
\docType{data}
\title{Capture-C counts of EScells and Neurons}
\description{Capture-C counts of EScells and Neurons restricted to a
  region of interest around the Paupar viewpoint.  This contains the
  same information as miniSE, but here packaged as a data frame.  Used
  in package documentation. }
\usage{miniSEDF}
\format{A data frame with 1909 rows.}
\keyword{datasets}