
\name{numPermutations}
\alias{numPermutations}
\docType{data}
\title{Number of permutations used in example permutation
  testing.}
\description{Number of permutations used in example permutation
  testing in the example in the documentation.}
\usage{numPermutations}
\format{A numeric scalar}
\keyword{datasets}
