% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaCaptureC.R
\name{rebinToMultiple}
\alias{rebinToMultiple}
\title{Rebin a SummarizedExperiment to a multiple of its bin width}
\usage{
rebinToMultiple(se, multiple, deleteShort = FALSE)
}
\arguments{
\item{se}{a RangedSummarizedExperiment to be rebinned}

\item{multiple}{the factor by which to fatten the bins}

\item{deleteShort}{= FALSE when set to true if the final bin is short
it will be deleted}
}
\value{
a RangedSummarizedExperiment
}
\description{
This is a faster way of rebinning when the old bins are consecutive
and constant width and the new bins are to be a multiple of that width
}
\examples{
rebinnedSummarizedExperiment = rebinToMultiple(binnedSummarizedExperiment,10)
}
