
\name{regionOfInterest}
\alias{regionOfInterest}
\docType{data}
\title{Region of interest surrounding the viewpoint}
\description{This is the region 500kb up- and down-stream of the Paupar
  viewpoint, and defines the region on which we will look for
  statistically significan chromatin remodeling.  This is used in
  package documentation.}
\usage{regionOfInterest}
\format{A GRanges object with one segment.}
\keyword{datasets}