
\name{significanceType}
\alias{significanceType}
\docType{data}
\title{Type for testing significance}
\description{Significance testing can be carried out by maximum, minimum
  or maximum absolute value. This has the value abs thus choosing the
  last of these.  This is used in package documentation.}
\usage{significanceType}
\format{A scalar character}
\keyword{datasets}