
\name{significantRegions}
\alias{significantRegions}
\docType{data}
\title{Regions of significant remodeling in example data}
\description{A GRanges object giving the value of delta in each bin and
  the significance or lack thereof according to the various statistical
  tests.  This is used in package documentation.}
\usage{significantRegions}
\format{A GRanges object}
\keyword{datasets}