
\name{significantRegionsPlot}
\alias{significantRegionsPlot}
\docType{data}
\title{A plot of the significant regions in the sample data.}
\description{A ggplot object show the values of delta in the region of
  interest and the significant sub-regions.  This is used for
  documentation purposes.}
\usage{significantRegionsPlot}
\format{a ggplot object}
\keyword{datasets}