
\name{smallBins}
\alias{smallBins}
\docType{data}
\title{Small Bins}
\description{The deltaCaptureC package depends on rebinning mean
  normalized capture C data from a small bin size to a large bin
  size. These are the small bins.  They have a bin size of 1000 bp,
  i.e., smallBinSize,  and
  span the region of interest.  This is used in package documentation.}
\usage{smallBins}
\format{a GRanges object}
\keyword{datasets}