
\name{smallBinSize}
\alias{smallBinSize}
\docType{data}
\title{Small Bin Size}
\description{The deltaCaptureC package depends on rebinning mean
  normalized capture C data from a small bin size to a large bin size.
  This gives the size of the former.  This is used in package documentation.}
\usage{smallBinSize}
\format{a numeric scalar}
\keyword{datasets}