
\name{smallerDeltaSE}
\alias{smallerDeltaSE}
\docType{data}
\title{A subset of miniDeltaSE.}
\description{This is a subset of miniDeltaSE. It's here so that the example
  for binSummarizedExperiment will run more quickly.}
\usage{smallerDeltaSE}
\format{A RangedSummarizedExperiment.}
\keyword{datasets}