
\name{smallSetOfSmallBins}
\alias{smallSetOfSmallBins}
\docType{data}
\title{Small Bins}
\description{This is a subset of the bins in smallBins and is here so
  that the example for binSummarizedExperiment will run more quickly.} 
\usage{smallSetOfSmallBins}
\format{a GRanges object}
\keyword{datasets}