
\name{viewpointRegion}
\alias{viewpointRegion}
\docType{data}
\title{Region surrounding the viewpoint}
\description{This is the region 50kb up- and down-stream of the Paupar
  viewpoint.  This is the region in which counts strongly depend on
  distance from the viewpoint and will be treated distinctly from the
  region of interest.}
\usage{viewpointRegion}
\format{A GRanges object with one segment.}
\keyword{datasets}